+++ R/gg_contour.R
@@ -443,7 +443,7 @@ gg_contour <- function(
         position = position,
         alpha = alpha,
         col = pal,
-        fill = pal,
+        # fill = pal,
         ...
       ) +
       coord +