+++ R/gg_errorbar.R
@@ -460,7 +460,7 @@ gg_errorbar <- function(
         position = position,
         alpha = alpha,
         col = pal,
-        fill = pal,
+        # fill = pal,
         ...
       ) +
       coord +