+++ R/gg_line.R
@@ -370,7 +370,7 @@ gg_line <- function(
         position = position,
         alpha = alpha,
         col = pal,
-        fill = pal,
+        # fill = pal,
         ...
       ) +
       coord +