+++ R/gg_pointrange.R
@@ -461,7 +461,7 @@ gg_pointrange <- function(
         position = position,
         alpha = alpha,
         col = pal,
-        fill = pal,
+        # fill = pal,
         ...
       ) +
       coord +