+++ R/gg_polygon.R
@@ -88,7 +88,7 @@
 #'
 #' datapoly <- merge(values, positions, by = c("id"))
 #'
-#' datapoly %>%
+#' datapoly |>
 #'   gg_polygon(
 #'     x = x,
 #'     y = y,