+++ R/gg_segment.R
@@ -428,7 +428,7 @@ gg_segment <- function(
         position = position,
         alpha = alpha,
         col = pal,
-        fill = pal,
+        # fill = pal,
         ...
       ) +
       coord +