+++ R/gg_step.R
@@ -373,7 +373,7 @@ gg_step <- function(
         position = position,
         alpha = alpha,
         col = pal,
-        fill = pal,
+        # fill = pal,
         ...
       ) +
       coord +