+++ R/gg_text.R
@@ -391,7 +391,7 @@ gg_text <- function(
         position = position,
         alpha = alpha,
         col = pal,
-        fill = pal,
+        # fill = pal,
         ...
       ) +
       coord +