+++ NEWS.md
@@ -1,5 +1,6 @@
 # ggblanket 5.1.0900
 
+* Removed unnecessary messages.
 * Fixed bug where `col_breaks` was not working for a numeric `col`.
 * Fixed bug where `str_keep_seq` was not working for date, datetime or time class.
 * Fixed bug so that time class variables work.